visualisation/pyproject.toml
[tool.poetry]
name = "visualisation"
version = "0.1.0"
description = ""
authors = ["Christoffer Björkskog <christoffer.bjorkskog@novia.fi>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
beautifulsoup4 = "4.10.0"
requests = "2.26.0"
pytest = "*"
python-dotenv = "*"
pyppeteer = "*"
pyppeteer_stealth = "*"
pytest-asyncio = "*"
python-fasthtml = "*"
openai = "*"
pysqlite3 = "*"
sqlite-vec = "*"
datasets = "*"



umap-learn = "0.5.6"
numba = "0.55.2"
numpy = "1.21.0"
plotly = "*"
hdbscan = "^0.8.38.post1"
rerankers = {extras = ["rankllm"], version = "^0.6.0"}
fh-matplotlib = "^0.0.4"
fh-frankenui = {git = "https://github.com/AnswerDotAI/fh-frankenui.git"}
graphviz = "^0.20.3"
selenium = "^4.27.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
